requirements.txt
certifi==2018.4.16
nltk==3.4.5
numpy==1.17.3
progressbar2==3.39.2
pycrayon==0.5
slackclient==2.2.1
torch==1.0.1.post2
torchvision==0.2.2.post3
transforms3d==0.3.1
